+++ src/components/Navbar.vue
@@ -1,0 +1,12 @@
+<template>
+  <ul>
+    <li>
+      <a href="/"> Home </a>
+      <a href="/about"> About </a>
+    </li>
+  </ul>
+</template>
+
+<style lang="scss">
+  @import "../scss/navbar.scss";
+</style>