+++ src/pages/About.vue
@@ -1,5 +1,19 @@
-<template></template>
+<script>
+  import Navbar from "../components/Navbar.vue";
+</script>
+
+<template>
+  <Navbar />
+  <div class="container">
+    <h1>About</h1>
+    <p>
+      Hello, there this is my about page. But you're probably wondering what
+      this website is about? Well, to sum it up since I'm lazy it's a website
+      created for educational purposes like learning Vue3 :D
+    </p>
+  </div>
+</template>
 
 <style lang="scss">
-  @import "../scss/styles.scss";
+  @import "../scss/about.scss";
 </style>