+++ CHANGELOG.md
@@ -7,3 +7,6 @@
 - 2022 February 28th 6:04PM
 - [+changes]
 - Updated the layout of the changelog using <https://keepachangelog.com/en/1.0.0/>
+- 2022 March 1st 2:35PM
+- [+changelog, +colors, +styles, -tasks, +pages, +components, +tasks]
+- Added more colors, added styles, finished two tasks from the ToDo list, added the about page, navbar and added two more tasks.